pyproject.toml
[project]
name = "2025-08-31-chihlee-raspberry"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.2",
    "google-generativeai>=0.8.5",
]
